Add support for using a specific time
Adds the --when flag to the command which allows for adding a timestamp
to use when printing. It support parsing either a specific datetime
given in the RFC 3999 format, or a relative timestamp.
`github.com/olebedev/when` is used to parse anything that can't be
parsed as an RFC 3999 timestamp, which allows giving timestamp like "7
am yesterday", or "half an hour ago".